Generate binary test-vector files containing R and the decoder output
As with the *_S.bin files, these are easier to check against a list
of hashes than text files on operating systems which use CRLF line
endings.  (sha256sum -c does not support special handling for text
files.)

The decoder output file is for debugging and an extra layer of
verification that the decoder is working correctly (the R and decR
binary files should have identical contents).